+++ dwds/lib/src/debugging/debugger.dart
@@ -394,28 +394,28 @@ class Debugger extends Domain {
     return await inspector.jsCallFunctionOn(receiver, expression, args);
   }
 
+  // TODO(elliette): https://github.com/dart-lang/webdev/issues/1501 Re-enable
+  // after checking with Chrome team if there is a way to check if the Chrome
+  // DevTools is showing an overlay. Both cannot be shown at the same time:
+
   // Renders the paused at breakpoint overlay over the application.
-  // TODO(elliette): Detect whether or not Chrome DevTools is also open, and
-  // if so don't show this overlay. Chrome DevTools also adds an overlay above
-  // this one. Currently we just make sure our overlay message is shorter than
-  // Chrome DevTools message, so that our overlay is hidden.
-  void _showPausedOverlay() async {
-    if (_pausedOverlayVisible) return;
-    handleErrorIfPresent(await _remoteDebugger?.sendCommand('DOM.enable'));
-    handleErrorIfPresent(await _remoteDebugger?.sendCommand('Overlay.enable'));
-    handleErrorIfPresent(await _remoteDebugger
-        ?.sendCommand('Overlay.setPausedInDebuggerMessage', params: {
-      'message': 'Paused',
-    }));
-    _pausedOverlayVisible = true;
-  }
+  // void _showPausedOverlay() async {
+  //   if (_pausedOverlayVisible) return;
+  //   handleErrorIfPresent(await _remoteDebugger?.sendCommand('DOM.enable'));
+  //   handleErrorIfPresent(await _remoteDebugger?.sendCommand('Overlay.enable'));
+  //   handleErrorIfPresent(await _remoteDebugger
+  //       ?.sendCommand('Overlay.setPausedInDebuggerMessage', params: {
+  //     'message': 'Paused',
+  //   }));
+  //   _pausedOverlayVisible = true;
+  // }
 
   // Removes the paused at breakpoint overlay from the application.
-  void _hidePausedOverlay() async {
-    if (!_pausedOverlayVisible) return;
-    handleErrorIfPresent(await _remoteDebugger?.sendCommand('Overlay.disable'));
-    _pausedOverlayVisible = false;
-  }
+  // void _hidePausedOverlay() async {
+  //   if (!_pausedOverlayVisible) return;
+  //   handleErrorIfPresent(await _remoteDebugger?.sendCommand('Overlay.disable'));
+  //   _pausedOverlayVisible = false;
+  // }
 
   /// Calls the Chrome Runtime.getProperties API for the object with [objectId].
   ///