+++ dwds/CHANGELOG.md
@@ -2,6 +2,7 @@
 - Change wording of paused overlay from "Paused in Dart DevTools" to "Paused"
 - Allow sending back the Dart DevTools URL from DWDS instead of launching 
   Dart DevTools, to support embedding Dart DevTools in Chrome DevTools.
+- Temporarily disable the paused in debugger overlay.
 
 ## 12.1.0
 - Update _fe_analyzer_shared to version ^34.0.0.